Refact Resource and ResourceNavigation to use Swipeable
